Add AWS API Gateway helpers
Apparently in Python you have to import the function that you want to
use in order to use it.

SO DUMB (Rather, not a thing you have to do in Ruby).

I see the performance benefit (not having to load hundreds of functions
if you only need one or two), but it's an extra cognitive step that I
don't have to worry about in Ruby.