Add functions for authenticated requests
Add a convenience function for issuing OAuth-authenticated requests
to the TripIt API.

The test is weak sauce. It should check whether the request
has the correct headers, at a minimum. However, I didn't know how to
mock a `requests` object without mocking the _whole_ object, which I
don't want to do (since I still want to see what the headers are).

This is quite easy to do with Ruby and Rspec, but quite challenging to
do with pytest. (`rspec-expectations` naturally supports asserting that
a module method is called, allows partial stubbing/mocking of methods,
makes it easy to set up test doubles and spies, etc.)

One SO post suggested `pickle`ing a real response object and loading
that as one's mock. That's not ideal either because that would mean
having to do the dance with TripIt's API manually, which I didn't want
to do at 10:50PM.

It's possible that I'm going about this all wrong.