Docker for Mac sucks :(
I developed this on a Mac. Volume mount performance with Docker for Mac
is abysmal and cuts my battery life in half _easily_. This is a known
issue between MacOS and the built-in lightweight hypervisor that comes
with this software. I would also experience constant CPU utilization
while Docker was running.

I decided to go old-school and switch over to Docker Toolbox. This uses
Virtualbox as its hypervisor. These problems seemed to have gone away.
However, the side effect of this is not using volume mounts anymore and
building my Docker image before unit and integration tests.

This change wasn't as slow as I thought it would be since the only
change here is copying over the app directory.