[fix] KeyError on trips without primary_location
Some trips apparently do not report a primary location. This causes a
`KeyError` to get thrown during execution and the API Gateway to send an
`HTTP 500` in response.

This commit adds better handling for when this occurs.

It does _not_ fix the fact that the API Gateway should return a `HTTP
422` with a more descriptive error mesage when this occurs. That will be
done separately...hopefully.

Signed-off-by: Carlos Nunez <13461447+carlosonunez@users.noreply.github.com>