[enhancement] Improve trip end time calculation accuracy
Previously, summarized trips returned by this API aligned to the last
flight in the trip.

This isn't always the case, e.g. road trips with flight segments.

This commit normalizes trip end dates to the end of the trip unless
the last flight in the trip happens on the same day as the trip's end
date.

Examples
---------

- Trip ends on 2023-08-01 + No flights = Trip ends on 2023-08-01 00:00 UTC
- Trip ends on 2023-08-01 + Last flight ends on 2023-07-31 18:00 UTC = Trip ends on 2023-08-01 00:00 UTC
- Trip ends on 2023-08-01 + Last flight ends on 2023-08-01 10:00 UTC = Trip ends on 2023-08-01 10:00 UTC

Assumptions
-----------

- Trips can never have an end date earlier than the last flight segment
  in the trip. (TripIt will usually block you from creating these, but
  I'm sure I haven't seen everything...)

Signed-off-by: Carlos Nunez <13461447+carlosonunez@users.noreply.github.com>